requirements.txt
flask
flask-jwt-extended
bcrypt
python-dotenv
requests
beautifulsoup4
pandas
openpyxl
pytest
PyJWT